Integrate dmotion-xblock
